Import/Export feeds in admin, edit FeedManager

+++ src/MiamBundle/Controller/AdminController.php
@@ -7,6 +7,7 @@ use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
 
 use MiamBundle\Entity\Feed;
 
@@ -45,20 +46,22 @@ class AdminController extends MainController
     	$form->handleRequest($request);
 
     	if($form->isSubmitted() && $form->isValid()) {
-            $this->get('feed_manager')->getFeedForUrl($form->get('url')->getData());
+            $this->get('feed_manager')->getFeedForUrl($form->get('url')->getData(), true);
     	}
 
     	return $this->redirectToRoute('admin');
     }
 
-    public function ajaxParseFeedAction($id) {
+    public function ajaxParseFeedAction(Request $request, $id) {
         $success = false;
 
-        $feed = $this->getRepo("Feed")->find($id);
-        if($feed) {
-            $this->get('data_parsing')->parseFeed($feed);
+        if($request->isXmlHttpRequest() && $this->isLoggedAdmin()) {
+            $feed = $this->getRepo("Feed")->find($id);
+            if($feed) {
+                $this->get('data_parsing')->parseFeed($feed);
 
-            $success = true;
+                $success = true;
+            }
         }
 
         return new JsonResponse(array(
@@ -66,18 +69,88 @@ class AdminController extends MainController
         ));
     }
 
-    public function ajaxDeleteFeedAction($id) {
+    public function ajaxDeleteFeedAction(Request $request, $id) {
         $success = false;
 
-        $feed = $this->getRepo("Feed")->find($id);
-        if($feed) {
-            $this->get('feed_manager')->deleteFeed($feed);
+        if($request->isXmlHttpRequest() && $this->isLoggedAdmin()) {
+            $feed = $this->getRepo("Feed")->find($id);
+            if($feed) {
+                $this->get('feed_manager')->deleteFeed($feed);
 
-            $success = true;
+                $success = true;
+            }
         }
 
         return new JsonResponse(array(
             'success' => $success
         ));
+    }
+
+    public function exportFeedsAction() {
+        $opml = $this->renderView("MiamBundle:Default:export.opml.twig", array(
+            'what' => "feeds",
+            'feeds' => $this->getRepo("Feed")->findAll(),
+            'ownerName' => "Admin"
+        ));
+
+        $response = new Response($opml);
+        $response->headers->set('Content-Type', "application/xml+opml");
+        $response->headers->set('Content-Disposition', "attachment;filename=Miam_Admin_".date('Y-m-d').".opml");
+
+        return $response;
+
+    }
+
+    public function ajaxPopupImportFeedsAction(Request $request) {
+        $response = array(
+            "success" => false
+        );
+
+        if($request->isXmlHttpRequest() && $this->isLoggedAdmin()) {
+            $html = $this->renderView("MiamBundle:Admin:popup_feeds_import.html.twig");
+
+            $response["success"] = true;
+            $response["html"] = $html;
+        }
+
+        return new JsonResponse($response);
+    }
+
+    public function importFeedsAction(Request $request) {
+        if($this->isTokenValid('admin_feeds_import', $request->get('csrf_token'))) {
+            $opml = false;
+            if(extension_loaded('SimpleXML') && isset($_FILES["opml"]["tmp_name"])) {
+                try {
+                    $opml = simplexml_load_file($_FILES["opml"]["tmp_name"]);
+                } catch(\Exception $e) {
+                    $opml = false;
+                }
+            }
+
+            if($opml !== false) {
+                $countNew = 0;
+
+                foreach($opml->body->children() as $outline) {
+                    if(isset($outline["xmlUrl"])) {
+                        $url = $outline["xmlUrl"];
+                        if(filter_var($url, FILTER_VALIDATE_URL) !== false) {
+                            $feed = $this->get("feed_manager")->findFeedForUrl($url);
+                            if(!$feed) {
+                                $feed = $this->get('feed_manager')->createFeedForUrl($url);
+                                $countNew++;
+                            }
+                        }
+                    }
+                }
+
+                $this->addFm("Feeds imported (".$countNew.")", "success");
+            } else {
+                $this->addFm("Error", "error");
+            }
+        } else {
+            $this->addFm("Invalid token", "error");
+        }
+
+        return $this->redirectToRoute("admin");
     }
 }
